Refactoring core training and viewing logic
