--- maddpg/requirements.txt
@@ -1,12 +1,0 @@
-agilerl>=0.1.16
-pettingzoo[classic,atari,mpe]>=1.23.1
-SuperSuit>=3.9.0
-torch>=2.0.1
-numpy>=1.24.2
-tqdm>=4.65.0
-fastrand==1.3.0
-gymnasium>=0.28.1
-imageio>=2.31.1
-Pillow>=9.5.0
-PyYAML>=5.4.1
-wandb>=0.13.10
